First version of the new Session class
get and get_custom can retrieve multiple datasets,  which are now held in a dictionary, so multiple can be stored in a single Session. Transformations can be applied on one, multiple or all datasets already retrieved. Transformations can also be tuned in order to apply different argument values to different datasets.